[Refactor] Refactor the components in articlePage and common directories
