+++ public/index.html
@@ -3,6 +3,10 @@
   <head>
     <meta charset="utf-8" />
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/earlyaccess/cwtexyen.css"
+    />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>ShengShi</title>
   </head>
