+++ src/components/common/HomepageBackground.jsx
@@ -4,19 +4,12 @@ import styled from 'styled-components';
 const HomepageBackground = () => {
   return (
     <BannerBG>
-      <div className="area">
-        <ul className="circles">
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-          <li></li>
-        </ul>
+      <div>
+        <Circles>
+          {[...Array(10).keys()].map((num) => (
+            <li key={num} />
+          ))}
+        </Circles>
       </div>
     </BannerBG>
   );
@@ -24,16 +17,17 @@ const HomepageBackground = () => {
 
 const BannerBG = styled.div`
   z-index: -100;
-  .circles {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-  }
+`;
 
-  .circles li {
+const Circles = styled.ul`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+
+  & li {
     position: absolute;
     display: block;
     list-style: none;
@@ -44,14 +38,14 @@ const BannerBG = styled.div`
     bottom: -150px;
   }
 
-  .circles li:nth-child(1) {
+  & li:nth-child(1) {
     left: 25%;
     width: 80px;
     height: 80px;
     animation-delay: 0s;
   }
 
-  .circles li:nth-child(2) {
+  & li:nth-child(2) {
     left: 10%;
     width: 20px;
     height: 20px;
@@ -59,14 +53,14 @@ const BannerBG = styled.div`
     animation-duration: 12s;
   }
 
-  .circles li:nth-child(3) {
+  & li:nth-child(3) {
     left: 70%;
     width: 20px;
     height: 20px;
     animation-delay: 4s;
   }
 
-  .circles li:nth-child(4) {
+  & li:nth-child(4) {
     left: 40%;
     width: 60px;
     height: 60px;
@@ -74,28 +68,28 @@ const BannerBG = styled.div`
     animation-duration: 18s;
   }
 
-  .circles li:nth-child(5) {
+  & li:nth-child(5) {
     left: 65%;
     width: 20px;
     height: 20px;
     animation-delay: 0s;
   }
 
-  .circles li:nth-child(6) {
+  & li:nth-child(6) {
     left: 75%;
     width: 110px;
     height: 110px;
     animation-delay: 3s;
   }
 
-  .circles li:nth-child(7) {
+  & li:nth-child(7) {
     left: 35%;
     width: 150px;
     height: 150px;
     animation-delay: 7s;
   }
 
-  .circles li:nth-child(8) {
+  & li:nth-child(8) {
     left: 50%;
     width: 25px;
     height: 25px;
@@ -103,7 +97,7 @@ const BannerBG = styled.div`
     animation-duration: 45s;
   }
 
-  .circles li:nth-child(9) {
+  & li:nth-child(9) {
     left: 20%;
     width: 15px;
     height: 15px;
@@ -111,7 +105,7 @@ const BannerBG = styled.div`
     animation-duration: 35s;
   }
 
-  .circles li:nth-child(10) {
+  & li:nth-child(10) {
     left: 85%;
     width: 150px;
     height: 150px;
