+++ src/components/common/LocationMap.jsx
@@ -1,3 +1,4 @@
+import styled from 'styled-components';
 import { compose, withProps } from 'recompose';
 import {
   withScriptjs,
@@ -8,47 +9,49 @@ import {
 
 require('dotenv').config();
 const GoogleMapsAPI = process.env.REACT_APP_GOOGLE_API_KEY;
+const googleMapURL = `https://maps.googleapis.com/maps/api/js?v=3.exp&libraries=geometry,drawing,places&key=${GoogleMapsAPI}`;
+
+const StyledLoadingElement = styled.div`
+  height: 100%;
+`;
+
+const StyledContainerElement = styled.div`
+  height: 200px;
+`;
+
+const StyledMapElement = styled.div`
+  height: 100%;
+  border-radius: 10px;
+`;
 
 const LocationMap = compose(
   withProps({
-    googleMapURL: `https://maps.googleapis.com/maps/api/js?v=3.exp&libraries=geometry,drawing,places&key=${GoogleMapsAPI}`,
-    loadingElement: (
-      <div
-        style={{
-          height: `100%`,
-        }}
-      />
-    ),
-    containerElement: (
-      <div
-        style={{
-          height: `200px`,
-        }}
-      />
-    ),
-    mapElement: <div style={{ height: `100%`, borderRadius: `10px` }} />,
+    googleMapURL,
+    loadingElement: <StyledLoadingElement />,
+    containerElement: <StyledContainerElement />,
+    mapElement: <StyledMapElement />,
     isMarkerShown: true,
   }),
   withScriptjs,
   withGoogleMap
-)((props) => (
+)(({ exchangeLocation }) => (
   <GoogleMap
     defaultZoom={12}
     defaultCenter={
-      props.exchangeLocation
+      exchangeLocation
         ? {
-            lat: props.exchangeLocation._lat,
-            lng: props.exchangeLocation._long,
+            lat: exchangeLocation._lat,
+            lng: exchangeLocation._long,
           }
         : { lat: 25.04267234987771, lng: 121.56497334150076 }
     }
   >
     <Marker
       position={
-        props.exchangeLocation
+        exchangeLocation
           ? {
-              lat: props.exchangeLocation._lat,
-              lng: props.exchangeLocation._long,
+              lat: exchangeLocation._lat,
+              lng: exchangeLocation._long,
             }
           : { lat: 25.04267234987771, lng: 121.56497334150076 }
       }
