+++ src/components/common/MyQRcode.jsx
@@ -1,8 +1,12 @@
 const QRCode = require('qrcode.react');
 
 const MyQRcode = ({ info }) => {
+  const value = `https://shengshi.itcosmos.co/personal/${info}`;
+  const src =
+    'https://firebasestorage.googleapis.com/v0/b/shengshi-8bc48.appspot.com/o/images%2Flogo%2Fshengshi-logo2.svg?alt=media&token=218bb2b8-ebaf-4f8d-809a-50912c6d2a6a';
+
   const initialProps = {
-    value: `https://shengshi.itcosmos.co/personal/${info}`,
+    value,
     size: 200,
     bgColor: '#ffffff',
     fgColor: '#000000',
@@ -10,7 +14,7 @@ const MyQRcode = ({ info }) => {
     includeMargin: false,
     renderAs: 'svg',
     imageSettings: {
-      src: 'https://firebasestorage.googleapis.com/v0/b/shengshi-8bc48.appspot.com/o/images%2Flogo%2Fshengshi-logo2.svg?alt=media&token=218bb2b8-ebaf-4f8d-809a-50912c6d2a6a',
+      src,
       x: null,
       y: null,
       height: 24,
