+++ src/components/common/ShareCardTag.jsx
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
 
 const ShareCardTag = styled.span`
+  padding: 5px 10px;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 16px;
   color: rgb(129, 129, 129);
-  box-shadow: 0 2px 6px 0 hsla(0, 0%, 0%, 0.2);
   border-radius: 5px;
-  padding: 5px 10px;
+  box-shadow: 0 2px 6px 0 hsla(0, 0%, 0%, 0.2);
 `;
 
 export default ShareCardTag;
