+++ src/components/common/comment/Comment.jsx
@@ -1,20 +1,18 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-
+import Ripples from 'react-ripples';
+import { ErrorMessage, Info, Message } from '../ErrorMessageUnits';
 import {
   handleConfirmCommentEdit,
   handleDeleteComment,
 } from '../../../utils/firebase';
-
-import Ripples from 'react-ripples';
-
-import { ErrorMessage, Info, Message } from '../ErrorMessageUnits';
 
 const Comment = ({ currentUser, share, comment, userData }) => {
   const [isEdit, setIsEdit] = useState(false);
   const [editedComment, setEditedComment] = useState('');
   const [errorMessage, setErrorMessage] = useState('');
   const [showErrorMessage, setShowErrorMessage] = useState(false);
+  const { author, createdAt, commentContent } = comment;
 
   const handleEnter = (e) => {
     if (e.charCode === 13) {
@@ -33,16 +31,16 @@ const Comment = ({ currentUser, share, comment, userData }) => {
   return (
     <>
       <CommentContainer>
-        <CommentAvatar src={comment.author.imageUrl} />
+        <CommentAvatar src={author.imageUrl} />
         <CommentContent>
           <CommentRow>
-            <CommentAuthor>{comment.author.displayName}</CommentAuthor>
+            <CommentAuthor>{author.displayName}</CommentAuthor>
             <CommentDateTime>
-              {comment.createdAt.toDate().toLocaleString()}
+              {createdAt.toDate().toLocaleString()}
             </CommentDateTime>
           </CommentRow>
           <CommentText
-            value={isEdit ? editedComment : comment.commentContent}
+            value={isEdit ? editedComment : commentContent}
             onChange={
               isEdit ? (e) => setEditedComment(e.target.value) : () => {}
             }
@@ -58,7 +56,7 @@ const Comment = ({ currentUser, share, comment, userData }) => {
       </CommentContainer>
 
       <CommentButtonRow>
-        {currentUser.uid === comment.author.id && (
+        {currentUser.uid === author.id && (
           <>
             <EditButton isEdit={isEdit} onClick={() => setIsEdit(true)}>
               編輯
@@ -101,18 +99,18 @@ const CommentContainer = styled.div`
 const CommentAvatar = styled.img`
   width: 30px;
   height: 30px;
-  object-fit: cover;
-  border-radius: 50%;
   margin-right: 10px;
   margin-top: 5px;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
+  object-fit: cover;
+  border-radius: 50%;
 `;
 
 const CommentContent = styled.div`
   flex-grow: 1;
+  padding: 8px;
   border: 1px solid lightskyblue;
   border-radius: 5px;
-  padding: 8px;
 `;
 
 const CommentRow = styled.div`
@@ -123,22 +121,22 @@ const CommentRow = styled.div`
 `;
 
 const CommentAuthor = styled.div`
+  margin-right: 10px;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 18px;
-  margin-right: 10px;
 `;
 
 const CommentDateTime = styled.div`
-  color: rgba(129, 129, 129, 0.68);
   font-size: 10px;
+  color: rgba(129, 129, 129, 0.68);
 `;
 
 const CommentText = styled.input`
-  outline: none;
-  border: ${(props) => (props.isEdit ? '1px solid #52b78854' : 'none')};
-  border-radius: 5px;
   width: 100%;
   padding: 5px 5px;
+  outline: none;
+  border: ${({ isEdit }) => (isEdit ? '1px solid #52b78854' : 'none')};
+  border-radius: 5px;
 `;
 
 const CommentButtonRow = styled.div`
@@ -149,35 +147,35 @@ const CommentButtonRow = styled.div`
 `;
 
 const StyledButton = styled.button`
+  padding: 3px 6px;
+  border-radius: 5px;
+  box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
   font-family: 'cwTeXYen', sans-serif;
   font-size: 16px;
-  border-radius: 5px;
-  padding: 3px 6px;
   cursor: pointer;
-  box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
 `;
 
 const EditButton = styled(StyledButton)`
-  display: ${(props) => (props.isEdit ? 'none' : 'block')};
-  color: white;
-  background-color: #52b788;
   margin-right: 5px;
+  display: ${({ isEdit }) => (isEdit ? 'none' : 'block')};
+  background-color: #52b788;
+  color: white;
 `;
 
 const ConfirmBtnRipples = styled(Ripples)`
-  border-radius: 5px;
   margin-right: 5px;
+  border-radius: 5px;
 `;
 
 const ConfirmButton = styled(StyledButton)`
-  display: ${(props) => (props.isEdit ? 'block' : 'none')};
-  color: white;
+  display: ${({ isEdit }) => (isEdit ? 'block' : 'none')};
   background-color: #1e88e5;
+  color: white;
 `;
 
 const DeleteButton = styled(StyledButton)`
-  color: #52b788;
   border: 1px solid #52b788;
+  color: #52b788;
 `;
 
 export default Comment;
