+++ src/components/common/popup/PopupUnits.jsx
@@ -1,18 +1,18 @@
-import { DialogOverlay, DialogContent } from '@reach/dialog';
 import styled from 'styled-components';
+import Ripples from 'react-ripples';
+import { DialogOverlay, DialogContent } from '@reach/dialog';
 import { AiFillCloseCircle } from 'react-icons/ai';
 import { HiLocationMarker } from 'react-icons/hi';
 import { BsCalendarCheckFill, BsFillInfoCircleFill } from 'react-icons/bs';
 import SaladImg from '../../../images/common/salad.svg';
-import Ripples from 'react-ripples';
 
 export const StyledDialogContent = styled(DialogContent)`
-  position: relative;
   width: 80vw;
   max-width: 800px;
   margin-top: 100px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  position: relative;
   border-radius: 10px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
 `;
 
 export const AlertDialogContent = styled(StyledDialogContent)`
@@ -20,26 +20,26 @@ export const AlertDialogContent = styled(StyledDialogContent)`
 `;
 
 export const PopClose = styled(AiFillCloseCircle)`
-  fill: #1e88e582;
-  border-radius: 50%;
-  opacity: 0.8;
-  cursor: pointer;
   position: absolute;
   top: 10px;
   right: 10px;
   width: 22px;
   height: 22px;
+  fill: #1e88e582;
+  border-radius: 50%;
+  opacity: 0.8;
+  cursor: pointer;
 `;
 
 export const PopTitleContainer = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
+  margin-bottom: 10px;
+  margin-top: 10px;
   padding: 10px;
   background-color: #52b788;
   border-radius: 10px;
-  margin-bottom: 10px;
-  margin-top: 10px;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
 `;
 
@@ -76,13 +76,13 @@ export const PopRow = styled.div`
 `;
 
 export const StyledLabel = styled.label`
-  background-color: #52b788aa;
   width: fit-content;
   padding: 5px 10px;
   border-radius: 5px;
-  color: white;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 18px;
+  color: white;
+  background-color: #52b788aa;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
 
   @media screen and (min-width: 600px) {
@@ -92,15 +92,15 @@ export const StyledLabel = styled.label`
 
 export const StyledInput = styled.input`
   flex-grow: 1;
-  border: none;
-  background: none;
-  outline: none;
-  border-bottom: 2px solid #d9d7d7;
-  padding: 5px 8px;
   margin-left: 10px;
   margin-top: 10px;
+  padding: 5px 8px;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 14px;
+  outline: none;
+  background: none;
+  border: none;
+  border-bottom: 2px solid #d9d7d7;
 
   @media screen and (min-width: 600px) {
     font-size: 18px;
@@ -109,15 +109,15 @@ export const StyledInput = styled.input`
 
 export const StyledSpan = styled.span`
   flex-grow: 1;
-  border: none;
-  background: none;
-  outline: none;
-  padding: 5px 8px;
   margin-left: 10px;
   margin-top: 10px;
+  padding: 5px 8px;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 14px;
   line-height: 24px;
+  border: none;
+  background: none;
+  outline: none;
 
   @media screen and (min-width: 600px) {
     font-size: 18px;
@@ -175,14 +175,14 @@ export const ImgUpload = styled.label`
   justify-content: center;
   align-items: center;
   margin-right: 5px;
-  border-radius: 5px;
   padding: 5px 10px;
-  cursor: pointer;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 18px;
-  background-color: lightskyblue;
   color: white;
+  background-color: lightskyblue;
+  border-radius: 5px;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
+  cursor: pointer;
 
   @media screen and (min-width: 600px) {
     font-size: 22px;
@@ -193,13 +193,12 @@ const StyledBtn = styled.button`
   display: flex;
   justify-content: center;
   align-items: center;
-  /* margin-right: 5px; */
-  border-radius: 5px;
   padding: 5px 10px;
-  cursor: pointer;
   font-family: 'cwTeXYen', sans-serif;
   font-size: 18px;
+  border-radius: 5px;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
+  cursor: pointer;
 
   @media screen and (min-width: 600px) {
     font-size: 22px;
