+++ src/components/common/search/SearchUnits.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { IoIosSearch } from 'react-icons/io';
 import Ripples from 'react-ripples';
+import { IoIosSearch } from 'react-icons/io';
 
 export const SearchContent = styled.div`
   display: flex;
@@ -20,13 +20,13 @@ export const SearchContent = styled.div`
 
 export const SearchOutline = styled.div`
   position: relative;
-  background: #b7e4c7;
   height: 60px;
-  border-radius: 40px;
+  margin-right: 1rem;
   padding: 10px;
   color: white;
-  margin-right: 1rem;
+  border-radius: 40px;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
+  background: #b7e4c7;
   backdrop-filter: blur(5px);
 
   @media screen and (max-width: 520px) {
@@ -35,44 +35,44 @@ export const SearchOutline = styled.div`
 `;
 
 export const SearchBar = styled.input`
-  font-family: 'cwTeXYen', sans-serif;
-  border: none;
-  background: none;
-  outline: none;
-  float: left;
   width: 240px;
   padding: 0 6px;
+  float: left;
+  background: none;
+  border: none;
+  outline: none;
+  font-family: 'cwTeXYen', sans-serif;
   font-size: 24px;
-  transition: 0.4s;
-  line-height: 40px;
   color: #40916c;
+  line-height: 40px;
+  transition: 0.4s;
 `;
 
 export const SearchIconContainer = styled.span`
-  float: right;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
   display: flex;
   justify-content: center;
   align-items: center;
+  width: 40px;
+  height: 40px;
+  float: right;
+  background: white;
+  border-radius: 50%;
   text-decoration: none;
   transition: 0.4;
-  background: white;
   cursor: pointer;
 `;
 
 export const SearchIcon = styled(IoIosSearch)`
-  fill: #1b4332;
   width: 20px;
   height: 20px;
+  fill: #1b4332;
 `;
 
 export const StyledRipples = styled(Ripples)`
   height: 60px;
+  color: #52b788;
   border-radius: 20px;
   border: 1px solid #a0a0968a;
-  color: #52b788;
   box-shadow: 0px 2px 6px 0px hsla(0, 0%, 0%, 0.2);
   backdrop-filter: blur(5px);
 `;
