+++ src/components/qrcodeConfirmPage/QRcodeConfirmPage.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState, useCallback } from 'react';
-import styled from 'styled-components';
 import { useParams, useHistory } from 'react-router-dom';
 import Title from '../personalPage/Title';
