+++ src/index.css
@@ -1,7 +1,3 @@
-/* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100;300;400;500;700;900&display=swap');
 @import url('https://fonts.googleapis.com/earlyaccess/cwtexyen.css');
 
@@ -92,7 +88,7 @@ video {
   font-size: 100%;
   vertical-align: baseline;
 }
-/* HTML5 display-role reset for older browsers */
+
 article,
 aside,
 details,
@@ -136,7 +132,6 @@ table {
 * {
   box-sizing: border-box;
   font: inherit;
-  /* font-family: 'Noto Sans TC', sans-serif; */
 }
 a {
   color: inherit;
