+++ src/index.js
@@ -13,12 +13,12 @@ import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 
 ReactDOM.render(
-  // <React.StrictMode>
-  <Router>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </Router>,
-  // </React.StrictMode>
+  <React.StrictMode>
+    <Router>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </Router>
+  </React.StrictMode>,
   document.getElementById('root')
 );
